unpacked 08042014 2_00_00_00 msp430-gcc-14r1-10-source.tar.bz2
18827a77f0c572890159d95aa5c79d23  # md5
93ebed10bf4cb9bd0dcae247d1b5f2c9d05c88a8fca5d86a6bb01aa1dcb45534  # sha256
http://software-dl.ti.com/msp430/msp430_public_sw/mcu/msp430/MSPGCC/latest/exports/msp430-gcc-14r1-10-source.tar.bz2

+++ tools/gcc/testsuite/g++.dg/torture/pr37146-1.C
@@ -1,36 +1,37 @@
 // PR c++/37146
 // { dg-do run }
+// { dg-require-effective-target int32plus }
 
 extern "C" void abort ();
-long int a, b;
-struct A { long int i:8; long int j:8; long int k:16; long int l:32; } c;
+int a, b;
+struct A { int i:8; int j:8; int k:16; int l:32; } c;
 
 void
-f1 (int x, long int y)
+f1 (int x, int y)
 {
   (x ? a : b) = y;
 }
 
 void
-f2 (int x, long int y)
+f2 (int x, int y)
 {
   (x ? c.i : c.j) = y;
 }
 
 void
-f3 (int x, long int y)
+f3 (int x, int y)
 {
   (x ? c.i : a) = y;
 }
 
 void
-f4 (int x, long int y)
+f4 (int x, int y)
 {
   (x ? c.i : c.k) = y;
 }
 
 void
-f5 (int x, long int y)
+f5 (int x, int y)
 {
   (x ? c.l : b) = y;
 }
