+++ tools/gcc/testsuite/gcc.c-torture/compile/20001205-1.c
@@ -2,7 +2,7 @@
 
 static inline unsigned long rdfpcr(void)
 {
-        unsigned long tmp, ret;
+        unsigned int tmp, ret;
         __asm__ ("" : "=r"(tmp), "=r"(ret));
         return ret;
 }
